+++ packages/web-core/CHANGELOG.md
@@ -1,3 +1,10 @@
+## [Unreleased]
+
+### Changed
+
+- Add tasks event support
+- Allow to manage entity items
+
 ## [2.7.1] - 2024-06-27
 
 ### Fixed